+++ src/views/toolbar-items/defined-navigation/defined-navigation-targets.ts
@@ -1,1 +1,0 @@
-export const DEFINED_NAVIGATION_TARGETS = ["day", "week", "month", "quarter", "year", "custom", "active"] as const;